+++ src/nti/webhooks/_schema.py
@@ -7,13 +7,17 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import re
 
 from zope.configuration.fields import GlobalInterface
 from zope.schema import InterfaceField
+from zope.schema import Id
 from zope.schema.interfaces import NotAnInterface
 from zope.schema.interfaces import InvalidURI
+from zope.schema.interfaces import InvalidId
 
 from zope.interface.interfaces import IObjectEvent
+from zope.principalregistry.metadirectives import TextId
 
 from nti.schema.field import ValidURI
 
@@ -65,3 +69,42 @@ class HTTPSURL(ValidURI):
         super(HTTPSURL, self)._validate(value)
         if not value.lower().startswith('https://'):
             raise InvalidURI(value).with_field_and_value(self, value)
+
+class PermissivePrincipalId(TextId):
+    """
+    A principal ID that allows more than just URIs and dotted names.
+
+    Ids must be text:
+
+        >>> PermissivePrincipalId().validate(b'abc')
+        Traceback (most recent call last):
+        ...
+        WrongType:...
+
+    They cannot contain whitespace::
+
+        >>> PermissivePrincipalId().validate(u'internal space')
+        Traceback (most recent call last):
+        ...
+        InvalidId: Whitespace is not allowed in an ID
+        >>> PermissivePrincipalId().validate(u'internal\ttab')
+        Traceback (most recent call last):
+        ...
+        InvalidId: Whitespace is not allowed in an ID
+
+    Email-addresses are allowed::
+
+        >>> PermissivePrincipalId().validate(u'sjohnson@riskmetrics.com')
+
+    As are plain names::
+
+        >>> PermissivePrincipalId().validate(u'sjohnson')
+
+    """
+
+    def _validate(self, value):
+        # We bypass most of the validation of the super class.
+        super(Id, self)._validate(value) # pylint:disable=bad-super-call
+        if re.search(r'\s', value, re.UNICODE):
+            # TODO: I18N
+            raise InvalidId('Whitespace is not allowed in an ID').with_field_and_value(self, value)