+++ CHANGES.rst
@@ -7,6 +7,8 @@
 
 - Add support for Python 3.9.
 
+- Principal IDs are no longer required to be URIs or dotted names. See
+  `issue 21 <https://github.com/NextThought/nti.webhooks/issues/21>`_.
 
 0.0.4 (2020-09-16)
 ==================